Avoid Framework dependency of the library, move to executable
Root has problem to generate a dictionary for the code depending on the include/boost/container/...
coming via Framework ... MemoryResources.h dependence.
